requirements.txt
flask>=3.0.2
werkzeug>=3.0.1
pillow>=10.2.0
diffusers>=0.26.3